pyproject.toml
[project]
name = "wikidatasearch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "astrapy>=2.0.1",
    "fastapi>=0.116.1",
    "fastapi-cache2>=0.2.2",
    "gradio>=5.49.1",
    "gunicorn>=23.0.0",
    "langchain-astradb==0.6.0",
    "pandas>=2.3.3",
    "plotly>=6.3.1",
    "pymysql>=1.1.2",
    "requests>=2.32.0",
    "setuptools>=80.9.0",
    "slowapi>=0.1.9",
    "sqlalchemy>=2.0.43",
    "stopwordsiso>=0.6.1",
    "uvicorn>=0.35.0",
]

[dependency-groups]
dev = [
    "ruff>=0.9.0"
]

[tool.ruff]
target-version = "py313"
line-length = 120

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # Pyflakes (catches undefined names, unused imports, etc.)
    "I",   # isort (import sorting)
    "D",   # pydocstring (function/class documentation)
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.isort]
known-first-party = [
    "wikidatasearch"
]